pyproject.toml
[tool.poetry]
name = "oz-04-main-001"
version = "0.1.0"
description = ""
authors = ["taewoo-dev <twin9774@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
django-storages = "^1.14.4"
boto3 = "^1.35.39"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"
gunicorn = "^23.0.0"
channels = "^4.1.0"
django-redis = "^5.4.0"
django-apscheduler = "^0.7.0"
pillow = "^11.0.0"
pytest-django = "^4.9.0"
djangorestframework = "^3.15.2"
channels-redis = "^4.2.0"
celery = "^5.4.0"
redis = "^5.1.1"
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.11.2"
django-stubs = "^5.1.0"
coverage = "^7.6.2"
django-extensions = "^3.2.3"
ipython = "^8.28.0"
djangorestframework-stubs = "^3.15.1"
django-debug-toolbar = "^4.4.6"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"
ignore_missing_imports = true

[tool.mypy.django]
ignore_missing_imports = true

[tool.mypy.rest_framework]
ignore_missing_imports = true

[tool.mypy.drf_spectacular]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "rest_framework.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework_simplejwt.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework.generics.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.coverage.report]
# 테스트 80퍼 이상이면 통과
fail_under = 80

[tool.coverage.run]
# 테스트 제외
omit = [
  ".mypy_cache/**/*",
  "*/migrations/*",
  "*/*test*.py",
  "manage.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
